+++ triangle/src/components/ClubNameBox.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: "Pretendard";
-    src: url("../../assets/fonts/Pretendard-Bold.subset.woff2") format("woff2");;
+    src: url("../assets/fonts/Pretendard-Bold.subset.woff2") format("woff2");;
 }
 
 html, body {
